System.Console.Haskeline.Internal

Documentation

debugTerminalKeys :: IO a

This function may be used to debug Haskeline's input.

It loops indefinitely; every time a key is pressed, it will print that key as it was recognized by Haskeline. Pressing Ctrl-C will stop the loop.

Haskeline's behavior may be modified by editing your ~/.haskeline file. For details, see: https://github.com/judah/haskeline/wiki/CustomKeyBindings